refactor(step2): 피드백 반영
