+++ src/main/kotlin/calculator/CalculatorApplication.kt
@@ -1,10 +1,9 @@
 package calculator
 
 import calculator.tasks.Calculator
-import calculator.util.Const.Companion.NOTICE_INPUT
+import calculator.ui.splitReadLine
 
 fun main() {
-    println(NOTICE_INPUT)
     with(Calculator()) {
         val anyOfCollection = splitReadLine()
         println(calculate(anyOfCollection))
