+++ src/main/kotlin/calculator/tasks/Calculator.kt
@@ -3,14 +3,9 @@ package calculator.tasks
 import calculator.data.Operand
 import calculator.data.Operator
 import calculator.util.extension.toDouble
-import calculator.util.Const.Companion.DELIMITERS_BLANK
-import calculator.util.Const.Companion.NOTICE_NOT_OPERAND_SYMBOL
-import calculator.util.extension.splitToCalculate
+import calculator.util.Message.Companion.NOTICE_NOT_OPERAND_SYMBOL
 
 class Calculator {
-    fun splitReadLine(): List<String> {
-        return readLine().splitToCalculate(DELIMITERS_BLANK)
-    }
 
     fun calculate(anyOfCollection: List<String>): String {
         return anyOfCollection.reduceIndexed { index, acc, element ->
